requirements.txt
matplotlib==3.8.3
numpy==1.26.4
pandas==2.2.2
speedtest-cli==2.1.3